requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.5
djangorestframework==3.14.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
MarkupSafe==2.1.3
msgpack==1.0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.5
websocket-client==1.6.3
zipp==3.17.0
zope.interface==6.0
